requirements.txt
#SwinIR
torch>=1.7
basicsr>=1.4.2
facexlib>=0.2.5
gfpgan>=1.3.7
realesrgan>=0.2.5
numpy
opencv-python
torchvision
scipy
tqdm
lmdb
pyyaml
yapf
timm
numpy
opencv-python-headless
Pillow
# Real EserGAn
setuptools
gradio
torchvision
addict
future
lmdb
pyyaml
requests
scikit-image
scipy
tb-nightly
tqdm
yapf
psutil
#GFPGAN
facexlib
gfpgan
realesrgan
opencv-python
scipy
#CodeFormer
codeformer-perceptor
